feat: totally remove compatibility mode and set it as default behavior to make things clearer
The Custom Classes plugin was using the "Compatibility mode" which can be enabled / disabled by
users from the settings tab. If that mode was enabled (default) the blocks targetted by a custom
class in Live Preview was rendered using the Read mode format which allows to maintain one CSS
snippets for both modes + target whole lists, tables and multiline code blocks. Considering that
without that the Custom Classes plugin is almost unusable, I've choosen to remove that compatibility
mode and set it as the default behavior.

BREAKING CHANGE: No more compatibility mode